% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_raster.R
\name{calcular_raster}
\alias{calcular_raster}
\title{Cálculo de variables ráster}
\usage{
calcular_raster(raster, area, dim, entorno = 0, nombre)
}
\arguments{
\item{raster}{Ráster original sobre el que se pretende realizar el cálculo.}

\item{area}{Área sobre la que se desea realizar el cálculo.}

\item{dim}{Tamaño de cada píxel del ráster resultante, en metros.}

\item{entorno}{Distancia sobre la que se pretende resumir la información del ráster original. Default = 0. En todo otro caso se calcula el promedio de los valores de los píxeles en un área circular con un radio igual a la distancia establecida en este parámetro.}

\item{nombre}{Nombre de la variable resultante. El ráster será guardado en el entorno de trabajo con éste nombre y la extensión ".tif".}
}
\value{
La función permite readecuar las resoluciones y proyecciones de diferentes rásters a las utilizadas en el proyecto. Además, el paŕametro "entorno" permite realizar un cálculo del promedio de la variable bajo análisis en el vecindario definido por el usuario.
}
\description{
Esta función devuelve un ráster que resume el cálculo realizado a partir de otro ráster, incluyendo fuentes de datos con diferente proyección y resolución espacial.
}
\examples{
library(terra)
original = rast("edificaciones.tif")
library(sf)
area_de_estudio = st_read("area.gpkg")
calcular_raster(original, area_de_estudio, dim = 50, entorno = 100, "prueba_raster")

}
\author{
Juan Pablo Carranza
}
