% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_vut.R
\name{calcular_vut}
\alias{calcular_vut}
\title{Homogeneización del precio de mercado de los inmuebles}
\usage{
calcular_vut(df, index, dependiente, independientes, dist)
}
\arguments{
\item{df}{Objeto 'sf' con geometría 'POINT'. Muestra georreferenciada con precios de la tierra cuyo valor se pretende expresar en términos homogéneos.}

\item{index}{Identificador único para cada variable.}

\item{dependiente}{Nombre de la variable dependiente, aquella que contiene el precio por metro cuadrado.}

\item{independientes}{Vector con los nombres de las columnas a utilizarse como variables independientes, y el valor de referencia correspondiente a un "lote típico". Si las variables son numéricas se debe indicar si el ajuste a un lote típico debe tomar como referencia la media o la mediana de dicha variable. En el caso de ser variables categóricas, se debe indicar el nombre de la categoría a tomarse como referencia. Ver ejemplo abajo.}

\item{dist}{Distancia que define el vecindario sobre el que se realizará la corrección por dependencia espacial".}
}
\value{
La función estima el precio por metro cuadrado homogeneizado, siguiendo a Carranza (2024). Toma como input un objeto espacial 'sf' y ajusta coeficientes de corrección según diferentes variables definidas por el usuario para reexpresar el precio en término de un 'lote típico' para el área de estudio. La función devuelve un objetivo llamado "resultado", que consiste en el df original al cual se agrega una columna correspondiente al valor homogeneizado (vut) y otra columna con el coeficiente de homogeneización aplicado (coef). Se informa, además, el parámetro de ajuste estimado por el modelo lineal para cada una de las variables independientes involucradas en el proceso de homogeneización.
}
\description{
Esta función estima el valor de las obervaciones muestrales expresado en términos homogéneos o comparables.
}
\examples{
# Cargamos los datos correspondientes a una muestra de 1000 terrenos.
load("data/terrenos.rda")

# Se define como "lote típico" a aquel que se corresponde a una operación de
# venta, realizada con escritura, y de una superficie igual a la mediana
# de los lotes de la ciudad.
# Se indica que la variable que informa sobre el precio por metro cuadrado
# es "valor_m2".
# Se define al vecindario en base a una distancia euclidiana de 500 metros.

calcular_vut(df = datos,
             index = "id",
             independientes = c("tipo_valor" = "Venta",
                                "superficie" = "mediana",
                                "situacion_juridica" = "Con Escritura"),
             dependiente = "valor_m2",
             dist = 500)

summary(resultado)

}
\author{
Juan Pablo Carranza
}
