% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{dat}
\alias{dat}
\title{Conjunto de datos `datos`}
\format{
Un data frame con las siguientes columnas:
\describe{
  \item{vut}{Valor de utilidad territorial del punto.}
  \item{geometry}{Coordenadas geográficas del punto.}
}
}
\source{
Datos geoespaciales procesados.
}
\usage{
dat
}
\description{
Este dataset contiene valores de utilidad territorial (`vut`) y su ubicación geográfica.
}
\keyword{datasets}
