% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminar_outliers.R
\name{eliminar_outliers}
\alias{eliminar_outliers}
\title{Identificación de valores espacialmente atípicos}
\usage{
eliminar_outliers(df, dist, variable, umbral, moran)
}
\arguments{
\item{df}{Dataframe sobre el que se realizará el cálculo. Debe ser un objeto de la librería 'sf' con geometría 'POINT'.}

\item{dist}{Distancia euclidiana que define la extensión del vecindario en el que se compararán los valores.}

\item{variable}{Nombre de la variable sobre la que se realizará el cálculo.}

\item{umbral}{Umbral para caracterizar a una observación como atípica en relación a lo observado en el vecindario. Por ejemplo, un umbral igual a 0.1 implicará que se caracterizarán como outliers a todas aquellas observaciones cuyo valor esté por encima o por debajo del observado en el vecindario en una magnitud igual a +/- 10\%.}

\item{moran}{Logical. Si asume el valor 'TRUE' se procede a identificar outliers mediante el cálculo del índice de Moran local. En este caso, el parámetro 'umbral' queda sin efecto.}
}
\value{
La función permite identificar valores que son muy diferentes a los observados en el entorno. Para ello, compara el valor de cada observación con el promedio del entorno (ponderado por la distancia, mediante la construcción de una matriz de pesos espaciales). También permite identificar outliers espaciales mediante el cálculo del índice de Moran local, siguiendo a Anselin (1995).
}
\description{
Esta función permite eliminar observaciones cuyo valor es muy diferente a los observados en el entorno. Permite identificar outliers espaciales mediante el cálculo del índice de Moran local, o mediante el promedio observado en el entorno ponderado por la distancia.
}
\examples{

# Cargamos la base de datos
load("resultado.rda")

# Se define la estimación de observaciones atípicas a partir de la variable 'vut'
# La distancia para el cálculo es de 50 metros lineales. Esto define la extensión del vecindario.
# El umbral para considerar a una observación como atípica se define igual a 0.1 (+/- 10\\%)
# No se realiza el cálculo mediante el índice de Moran local.
eliminar_outliers(df = resultado,
                  dist = 50,
                  variable = "vut",
                  umbral = 0.1,
                  moran = FALSE)

# Si, en cambio, se desea identificar valores atípicos mediante el cálculo del índice de Moran local:
eliminar_outliers(df = resultado,
                  dist = 50,
                  variable = "vut",
                  moran = TRUE)

}
\references{
Anselin, L. (1995). Local Indicators of Spatial Association—LISA.
\emph{Geographical Analysis}, \bold{27}(2), 93–115.
\doi{10.1111/j.1538-4632.1995.tb00338.x}. Disponible en:
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1538-4632.1995.tb00338.x}.
}
\author{
Juan Pablo Carranza
}
