% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{data}
\name{terrenos}
\alias{terrenos}
\title{Conjunto de datos de terrenos}
\format{
Un data frame con las siguientes columnas:
\describe{
  \item{id}{Identificador único del terreno.}
  \item{valor}{Valor total del terreno en la moneda especificada.}
  \item{moneda}{Moneda en la que está expresado el valor del terreno (por ejemplo, "Dólar (USD)").}
  \item{tipo_valor}{Tipo de valor registrado (por ejemplo, "Oferta").}
  \item{superficie}{Superficie total del terreno en metros cuadrados.}
  \item{situacion_juridica}{Estado legal del terreno (por ejemplo, "Con Escritura", "Pre venta").}
  \item{valor_m2}{Valor por metro cuadrado del terreno.}
  \item{geom}{Coordenadas del terreno en el sistema de referencia utilizado.}
}
}
\source{
Datos generados a partir de anuncios de terrenos en venta.
}
\usage{
terrenos
}
\description{
Este dataset contiene información sobre terrenos en venta, incluyendo el precio, moneda, tipo de valor, superficie y situación jurídica.
}
\keyword{datasets}
