% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_dist.R
\name{calcular_dist}
\alias{calcular_dist}
\title{Cálculo de variables de distancia}
\usage{
calcular_dist(area, objeto, dim, nombre)
}
\arguments{
\item{area}{Polígono que indica el área sobre la cuál se ejecutará el cálculo.}

\item{objeto}{Objeto espacial "sf" contra el que se calcularán las distancias.}

\item{dim}{Tamaño de cada píxel del ráster resultante, en metros.}

\item{nombre}{Nombre de la variable resultante. El ráster será guardado en el entorno de trabajo con éste nombre y la extensión ".tif".}
}
\value{
Se genera un ráster que informa, para cada píxel, sobre la distancia (en metros lineales) al objeto más cercano.
}
\description{
Esta función devuelve un ráster con las distancias euclidianas, en metros, desde cada píxel hacia el objeto vectorial más cercano.
}
\examples{
# Definimos el área de estudio
library(nominatimlite)
library(osmdata)
library(sf)
area_de_estudio <- geo_lite_sf(address = "La Plata, Argentina", points_only = F)
bbox = st_transform(area_de_estudio, 4326) |> st_bbox(bbox)

# Descargamos las vias principales
vias_prim <- opq(bbox) |>
  add_osm_feature(key = "highway", value = "secondary") |>
  osmdata_sf()
vias_prim <- vias_prim$osm_lines

# Calculamos un ráster con las distancias a la vía principal más cercana
calcular_dist(area = area_de_estudio,
              objeto = vias_prim,
              dim = 50,
              nombre = "vias")
}
\author{
Juan Pablo Carranza
}
