% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_raster.R
\name{calcular_raster}
\alias{calcular_raster}
\title{Calculo de variables raster}
\usage{
calcular_raster(raster, area, dim, entorno = 0, nombre)
}
\arguments{
\item{raster}{Raster original sobre el que se pretende realizar el calculo.}

\item{area}{Area sobre la que se desea realizar el calculo.}

\item{dim}{Tamano de cada pixel del raster resultante, en metros.}

\item{entorno}{Distancia sobre la que se pretende resumir la informacion del raster original. Default = 0. En todo otro caso se calcula el promedio de los valores de los pixeles en un area circular con un radio igual a la distancia establecida en este parametro.}

\item{nombre}{Nombre de la variable resultante. El raster sera guardado en el entorno de trabajo con este nombre y la extension ".tif".}
}
\value{
La funcion permite readecuar las resoluciones y proyecciones de diferentes rasters a las utilizadas en el proyecto. Ademas, el parametro "entorno" permite realizar un calculo del promedio de la variable bajo analisis en el vecindario definido por el usuario.
}
\description{
Esta funcion devuelve un raster que resume el calculo realizado a partir de otro raster, incluyendo fuentes de datos con diferente proyeccion y resolucion espacial.
}
\examples{
original = terra::rast("data/edificaciones.tif")
area_de_estudio = sf::st_read("inst/extdata/area.gpkg")
calcular_raster(original, area_de_estudio, dim = 50, entorno = 100, "prueba_raster")
}
\author{
Juan Pablo Carranza
}
