% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminar_outliers.R
\name{eliminar_outliers}
\alias{eliminar_outliers}
\title{Identificacion de valores espacialmente atipicos}
\usage{
eliminar_outliers(df, dist, variable, umbral, moran)
}
\arguments{
\item{df}{Dataframe sobre el que se realizara el calculo. Debe ser un objeto de la libreria 'sf' con geometria 'POINT'.}

\item{dist}{Distancia euclidiana que define la extension del vecindario en el que se compararan los valores.}

\item{variable}{Nombre de la variable sobre la que se realizara el calculo.}

\item{umbral}{Umbral para caracterizar a una observacion como atipica en relacion a lo observado en el vecindario. Por ejemplo, un umbral igual a 0.1 implicara que se caracterizaran como outliers a todas aquellas observaciones cuyo valor este por encima o por debajo del observado en el vecindario en una magnitud igual a +/- 10\%.}

\item{moran}{Logical. Si asume el valor 'TRUE' se procede a identificar outliers mediante el calculo del indice de Moran local. En este caso, el parametro 'umbral' queda sin efecto.}
}
\value{
La funcion permite identificar valores que son muy diferentes a los observados en el entorno. Para ello, compara el valor de cada observacion con el promedio del entorno (ponderado por la distancia, mediante la construccion de una matriz de pesos espaciales). Tambien permite identificar outliers espaciales mediante el calculo del indice de Moran local, siguiendo a Anselin (1995).
}
\description{
Esta funcion permite eliminar observaciones cuyo valor es muy diferente a los observados en el entorno. Permite identificar outliers espaciales mediante el calculo del indice de Moran local, o mediante el promedio observado en el entorno ponderado por la distancia.
}
\examples{

# Cargamos la base de datos
data("resultado", package = "valuate")

# Se define la estimacion de observaciones atipicas a partir de la variable 'vut'
# La distancia para el calculo es de 50 metros lineales. Esto define la extension del vecindario.
# El umbral para considerar a una observacion como atipica se define igual a 0.1 (+/- 10\\%)
# No se realiza el calculo mediante el indice de Moran local.
eliminar_outliers(df = resultado,
                  dist = 50,
                  variable = "vut",
                  umbral = 0.1,
                  moran = FALSE)

# Si, en cambio, se desea identificar valores atipicos mediante el calculo del indice de Moran local:
eliminar_outliers(df = resultado,
                  dist = 50,
                  variable = "vut",
                  moran = TRUE)
}
\references{
Anselin, L. (1995). Local Indicators of Spatial Association - LISA.
\emph{Geographical Analysis}, \bold{27}(2), 93-115.
\doi{10.1111/j.1538-4632.1995.tb00338.x}. Disponible en:
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1538-4632.1995.tb00338.x}.
}
\author{
Juan Pablo Carranza
}
